data_ingestion/requirements.txt
boto3
psycopg2-binary
openai
pinecone
sentence-transformers
python-dotenv